+++ src/app/workspace/twin/[twinId]/views/twin-history.tsx
@@ -1,8 +1,8 @@
 import { HistoryV1, useFetchHistoryV1 } from "@/entities/twin";
 import { User, UserResourceLink } from "@/entities/user";
 import { PagedResponse } from "@/shared/api";
-import { DataTableHandle, GuidWithCopy } from "@/shared/ui";
-import { Experimental_CrudDataTable } from "@/widgets/crud-data-table";
+import { GuidWithCopy } from "@/shared/ui";
+import { CrudDataTable, DataTableHandle } from "@/widgets/crud-data-table";
 import { ColumnDef, PaginationState } from "@tanstack/table-core";
 import { useContext, useRef } from "react";
 import { toast } from "sonner";
@@ -102,7 +102,7 @@ export function TwinHistory() {
 
   return (
     <div className="mb-10">
-      <Experimental_CrudDataTable
+      <CrudDataTable
         title="History"
         ref={tableRef}
         columns={[
