+++ src/app/workspace/twin/[twinId]/views/twin-links.tsx
@@ -4,9 +4,8 @@ import { TwinLinkView, useFetchTwinLinks } from "@/entities/twinLink";
 import { UserResourceLink } from "@/entities/user";
 import { PagedResponse } from "@/shared/api";
 import { formatToTwinfaceDate } from "@/shared/libs";
-import { DataTableHandle } from "@/shared/ui/data-table/data-table";
 import { GuidWithCopy } from "@/shared/ui/guid";
-import { Experimental_CrudDataTable } from "@/widgets/crud-data-table";
+import { CrudDataTable, DataTableHandle } from "@/widgets/crud-data-table";
 import { ColumnDef } from "@tanstack/table-core";
 import { useContext, useRef } from "react";
 import { toast } from "sonner";
@@ -98,7 +97,7 @@ export function TwinLinks() {
 
   return (
     <>
-      <Experimental_CrudDataTable
+      <CrudDataTable
         title="Forward Links"
         ref={tableRefForward}
         columns={[
@@ -120,7 +119,7 @@ export function TwinLinks() {
         ]}
       />
 
-      <Experimental_CrudDataTable
+      <CrudDataTable
         title="Backward Links"
         ref={tableRefBackward}
         columns={[
