+++ src/screens/permission/views/grant-section/users-table.tsx
@@ -9,7 +9,7 @@ import {
 import { PermissionContext } from "@/features/permission";
 import { PagedResponse } from "@/shared/api";
 import { isUndefined } from "@/shared/libs";
-import { Experimental_CrudDataTable } from "@/widgets/crud-data-table";
+import { CrudDataTable } from "@/widgets/crud-data-table";
 import { ColumnDef, PaginationState } from "@tanstack/react-table";
 import { useContext } from "react";
 import { toast } from "sonner";
@@ -85,7 +85,7 @@ export function UsersTable() {
   if (isUndefined(permission)) return null;
 
   return (
-    <Experimental_CrudDataTable
+    <CrudDataTable
       title="Users"
       columns={[
         colDefs.id,
