+++ src/screens/twinclassTriggers/twinflow-transition-triggers.tsx
@@ -7,10 +7,10 @@ import {
   useTwinFlowTransitionTriggersSearch,
 } from "@/entities/twinFlowTransition";
 import { useContext, useRef } from "react";
-import { DataTableHandle } from "@/shared/ui/data-table/data-table";
+import { DataTableHandle } from "@/widgets/crud-data-table";
 import { ColumnDef } from "@tanstack/table-core";
 import { GuidWithCopy } from "@/shared/ui/guid";
-import { Experimental_CrudDataTable } from "@/widgets/crud-data-table";
+import { CrudDataTable } from "@/widgets/crud-data-table";
 import { Check } from "lucide-react";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
@@ -137,7 +137,7 @@ export function TwinflowTransitionTriggers({
   }
 
   return (
-    <Experimental_CrudDataTable
+    <CrudDataTable
       ref={tableRef}
       className="mt-4"
       title="Triggers"
