--- src/shared/ui/data-table/crud-data-table.tsx
@@ -1,342 +1,0 @@
-"use client";
-
-import { AutoField, AutoFormValueInfo } from "@/components/auto-field";
-import { PagedResponse } from "@/shared/api";
-import { cn, fixedForwardRef } from "@/shared/libs";
-import { Button } from "@/shared/ui/button";
-import { CustomizableColumnsPopover } from "@/shared/ui/data-table/crud-data-table-columns-popover";
-import {
-  DataTable,
-  DataTableHandle,
-  DataTableProps,
-  DataTableRow,
-} from "@/shared/ui/data-table/data-table";
-import { Form } from "@/shared/ui/form";
-import { Input } from "@/shared/ui/input";
-import { Popover, PopoverContent, PopoverTrigger } from "@/shared/ui/popover";
-import { Separator } from "@/shared/ui/separator";
-import { PaginationState } from "@tanstack/table-core";
-import { FilterIcon, RefreshCw, Search } from "lucide-react";
-import {
-  ForwardedRef,
-  useEffect,
-  useImperativeHandle,
-  useMemo,
-  useRef,
-  useState,
-} from "react";
-import { DefaultValues, useForm } from "react-hook-form";
-
-export interface FiltersState<TFilterKeys extends string = string> {
-  search?: string;
-  filters: Record<TFilterKeys, any>;
-}
-
-interface CrudDataTableCreateButtonProps {
-  enabled?: boolean;
-  text?: string;
-  onClick?: () => any;
-}
-
-interface CrudDataTableSearchProps {
-  enabled?: boolean;
-  placeholder?: string;
-}
-
-interface CrudDataTableFiltersProps<TFilterKeys extends string = string> {
-  filtersInfo: Record<TFilterKeys, AutoFormValueInfo>;
-  onChange?: (values: Record<TFilterKeys, any>) => Promise<any>;
-}
-
-interface CrudDataTableCustomizableColumnsProps {
-  enabled?: boolean;
-  defaultVisibleKeys?: string[];
-}
-
-interface CrudDataTableProps<
-  TData extends DataTableRow<TData>,
-  TValue,
-  TFilterKeys extends string = string,
-> extends Omit<DataTableProps<TData, TValue>, "fetcher"> {
-  fetcher: (
-    pagination: PaginationState,
-    filters: FiltersState<TFilterKeys>
-  ) => Promise<PagedResponse<TData>>;
-  title?: string;
-  createButton?: CrudDataTableCreateButtonProps;
-  search?: CrudDataTableSearchProps;
-  filters?: CrudDataTableFiltersProps;
-  customizableColumns?: CrudDataTableCustomizableColumnsProps;
-  hideRefresh?: boolean;
-  className?: string;
-}
-
-/**
- * @deprecated
- * The `CrudDataTable` should not be used in new code.
- * Use `Experimental_CrudDataTable` instead.
- */
-export const CrudDataTable = fixedForwardRef(CrudDataTableInternal);
-
-function CrudDataTableInternal<
-  TData extends DataTableRow<TData>,
-  TValue,
-  TFilterKeys extends string,
->(
-  {
-    title,
-    createButton,
-    search,
-    filters,
-    customizableColumns,
-    hideRefresh,
-    className,
-    fetcher,
-    ...props
-  }: CrudDataTableProps<TData, TValue, TFilterKeys>,
-  ref: ForwardedRef<DataTableHandle>
-) {
-  const [tableSearch, setTableSearch] = useState<string>("");
-  const [tableFilters, setTableFilters] = useState<Record<TFilterKeys, any>>(
-    {} as Record<TFilterKeys, any>
-  );
-  const [visibleColumnKeys, setVisibleColumnKeys] = useState<string[]>(
-    customizableColumns?.defaultVisibleKeys ?? []
-  );
-  const [sortColumnKeys, setSortColumnKeys] = useState<string[]>(
-    props.columns?.map((col) => getColumnKey(col)) ?? []
-  );
-  const tableRef = useRef<DataTableHandle>(null);
-  useImperativeHandle(ref, () => tableRef.current!, [tableRef]);
-
-  function fetchWrapper(pagination: PaginationState) {
-    return fetcher(pagination, { search: tableSearch, filters: tableFilters });
-  }
-
-  async function refresh() {
-    tableRef.current?.refresh();
-  }
-
-  async function onFiltersChange(values: Record<TFilterKeys, any>) {
-    setTableFilters(values);
-    filters?.onChange?.(values);
-  }
-
-  function onVisibleColumnsChange(columns: string[]) {
-    setVisibleColumnKeys(columns);
-  }
-
-  function onSortColumnsChange(columns: string[]) {
-    setSortColumnKeys(columns);
-  }
-
-  useEffect(() => {
-    tableRef.current?.resetPage();
-    refresh();
-  }, [tableFilters]);
-
-  const visibleColumns = useMemo(() => {
-    return customizableColumns?.enabled
-      ? props.columns
-          .filter((column) => {
-            return visibleColumnKeys.includes(getColumnKey(column));
-          })
-          .sort((a, b) => {
-            return (
-              sortColumnKeys.indexOf(getColumnKey(a)) -
-              sortColumnKeys.indexOf(getColumnKey(b))
-            );
-          })
-      : props.columns;
-  }, [customizableColumns, visibleColumnKeys, sortColumnKeys]);
-
-  function getColumnKey(column: any) {
-    return column.accessorKey ? (column.accessorKey as string) : column.id!;
-  }
-
-  return (
-    <div className={cn("flex-1 py-4", className)}>
-      <div className="mb-2 flex justify-between">
-        <div className={"flex items-center"}>
-          {title && <div className="text-lg">{title}</div>}
-          {search?.enabled && (
-            <>
-              <form
-                className="flex flex-row space-x-1"
-                onSubmit={(e) => {
-                  e.preventDefault();
-                  tableRef.current?.refresh();
-                }}
-              >
-                <Input
-                  placeholder={search?.placeholder ?? "Search..."}
-                  value={tableSearch}
-                  onChange={(event) => setTableSearch(event.target.value)}
-                  className="max-w-sm"
-                />
-                <Button variant={"ghost"} type="submit">
-                  <Search />
-                </Button>
-              </form>
-            </>
-          )}
-        </div>
-        <div className={"flex space-x-4"}>
-          {!hideRefresh && (
-            <>
-              <Button variant="ghost" onClick={refresh}>
-                <RefreshCw />
-              </Button>
-            </>
-          )}
-          {filters && (
-            <FiltersPopover
-              filtersInfo={filters.filtersInfo}
-              onChange={onFiltersChange}
-            />
-          )}
-          {customizableColumns?.enabled && (
-            <CustomizableColumnsPopover
-              columns={props.columns
-                .map((column) => {
-                  var columnAsAny = column as any;
-                  if (columnAsAny.accessorKey) {
-                    return {
-                      id: columnAsAny.accessorKey as string,
-                      name: column.header as string,
-                      visible: visibleColumnKeys.includes(
-                        columnAsAny.accessorKey as string
-                      ),
-                    };
-                  } else {
-                    return {
-                      id: column.id as string,
-                      name: column.header as string,
-                      visible: visibleColumnKeys.includes(column.id as string),
-                    };
-                  }
-                })
-                .sort((a, b) => {
-                  return (
-                    sortColumnKeys.indexOf(a.id) - sortColumnKeys.indexOf(b.id)
-                  );
-                })}
-              sortKeys={sortColumnKeys}
-              onVisibleChange={onVisibleColumnsChange}
-              onSortChange={onSortColumnsChange}
-              onReset={() =>
-                setVisibleColumnKeys(
-                  customizableColumns.defaultVisibleKeys ?? []
-                )
-              }
-            />
-          )}
-          {(!hideRefresh || filters) && <Separator orientation={"vertical"} />}
-          {createButton?.enabled && (
-            <Button onClick={createButton?.onClick}>
-              {createButton?.text ?? "Create"}
-            </Button>
-          )}
-        </div>
-      </div>
-
-      <DataTable
-        ref={tableRef}
-        {...props}
-        columns={visibleColumns}
-        fetcher={fetchWrapper}
-      />
-    </div>
-  );
-}
-
-interface FiltersPopoverProps<TFilterKeys extends string = string> {
-  filtersInfo: Record<TFilterKeys, AutoFormValueInfo>;
-  onChange: (values: Record<TFilterKeys, any>) => Promise<any>;
-}
-
-function FiltersPopover<TFilterKeys extends string = string>({
-  filtersInfo,
-  onChange,
-}: FiltersPopoverProps<TFilterKeys>) {
-  const [open, setOpen] = useState(false);
-
-  const keys = Object.keys(filtersInfo) as TFilterKeys[];
-
-  const form = useForm({
-    defaultValues: Object.fromEntries(
-      keys.map((key) => [key, filtersInfo[key]!.defaultValue ?? ""])
-    ) as DefaultValues<Record<TFilterKeys, any>>,
-  });
-
-  async function internalSubmit(newValue: Record<TFilterKeys, any>) {
-    try {
-      await onChange(newValue);
-      setOpen(false);
-    } catch (e) {
-      console.error("Failed to update FiltersPopover", e);
-    }
-  }
-
-  async function onReset() {
-    form.reset(undefined, { keepDefaultValues: true });
-
-    try {
-      await onChange({} as Record<TFilterKeys, any>);
-      setOpen(false);
-    } catch (e) {
-      console.error("Failed to reset FiltersPopover", e);
-    }
-  }
-
-  return (
-    <Popover open={open} onOpenChange={setOpen}>
-      <PopoverTrigger asChild>
-        <Button
-          type="button"
-          className={cn("block" /*, className*/)}
-          // name={name}
-          onClick={() => {
-            setOpen(true);
-          }}
-          variant="default"
-        >
-          <FilterIcon />
-        </Button>
-      </PopoverTrigger>
-      <PopoverContent className="max-h-[80vh] overflow-y-auto">
-        <Form {...form}>
-          <form
-            onSubmit={form.handleSubmit(internalSubmit)}
-            className="space-y-4"
-          >
-            {keys.map((filterKey) => {
-              return (
-                <AutoField
-                  key={filterKey}
-                  info={filtersInfo[filterKey]!}
-                  name={filterKey}
-                  control={form.control}
-                />
-              );
-            })}
-
-            <div className={"flex flex-row justify-end gap-2"}>
-              <Button
-                onClick={onReset}
-                type="reset"
-                variant="outline"
-                loading={form.formState.isSubmitting}
-              >
-                Clear
-              </Button>
-              <Button type="submit" loading={form.formState.isSubmitting}>
-                Apply
-              </Button>
-            </div>
-          </form>
-        </Form>
-      </PopoverContent>
-    </Popover>
-  );
-}
