--- src/shared/ui/data-table/index.ts
@@ -1,4 +1,0 @@
-export * from "./crud-data-table";
-export * from "./crud-data-table-columns-popover";
-export * from "./data-table";
-export * from "./data-table-pagination";
