+++ src/widgets/crud-data-table/data-table/index.ts
@@ -1,0 +1,2 @@
+export * from "./data-table";
+export * from "./types";
