+++ src/widgets/crud-data-table/index.ts
@@ -1,1 +1,7 @@
-export * from "./experimental_crud-data-table";
+export * from "./crud-data-table";
+export type {
+  DataTableHandle,
+  DataTableProps,
+  DataTableRow,
+} from "./data-table";
+export * from "./types";
