+++ src/widgets/crud-data-table/types.ts
@@ -1,0 +1,4 @@
+export interface FiltersState<TFilterKeys extends string = string> {
+  search?: string;
+  filters: Record<TFilterKeys, any>;
+}
